requirements.txt
numpy
pandas
sklearn
plotly.express
json
geojson_rewind